# docs/swagger.yaml
definitions:
...
  handler.CreateProjectRequest:
    properties:
      name:
        example: My React App
        type: string
    type: object
...
  handler.DateCount:
    properties:
      count:
        example: 42
        type: integer
      date:
        example: "2024-01-15"
        type: string
    type: object
  handler.ErrorResponse:
    properties:
      error:
        example: invalid JSON
        type: string
    type: object
...
  handler.StatsResponse:
    properties:
      by_date:
        items:
          $ref: '#/definitions/handler.DateCount'
        type: array
      by_level:
        additionalProperties:
          format: int64
          type: integer
        type: object
      total_issues:
        example: 15
        type: integer
      total_occurrences:
        example: 59
        type: integer
    type: object
...
  handler.UploadSourcemapResponse:
    properties:
      filename:
        example: index-abc123.js.map
        type: string
      object_key:
        example: proj-uuid/a3f9c2b/index-abc123.js.map
        type: string
      release:
        example: a3f9c2b
        type: string
    type: object
...
  models.Project:
    properties:
      api_key:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/projects:
    get:
      description: Returns all projects visible to authenticated users (company-wide).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Project'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List projects
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: Creates a new project and generates an API key. The returned `id`
        is what the errhub-package SDK uses as `projectId`.
      parameters:
      - description: Project details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create project
      tags:
      - projects
  /api/projects/{projectId}:
    delete:
      description: Permanently deletes a project and all its issues and events (cascade).
      parameters:
      - description: Project UUID
        in: path
        name: projectId
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete project
      tags:
      - projects
    get:
      description: Returns a single project by ID.
      parameters:
      - description: Project UUID
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Project'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get project
      tags:
      - projects
...
  /api/projects/{projectId}/releases/{release}/sourcemaps:
    post:
      consumes:
      - multipart/form-data
      description: Uploads a `.map` file for a specific build release. Called automatically
        from CI after `npm run build`. The release string is typically the short git
        commit hash (`$GITHUB_SHA` or `git rev-parse --short HEAD`). Files are stored
        in MinIO and tracked in the DB for automatic retention management (deleted
        after 3 days of no use).
      parameters:
      - description: Project UUID
        in: path
        name: projectId
        required: true
        type: string
      - description: Build release string (e.g. git commit hash)
        in: path
        name: release
        required: true
        type: string
      - description: Source map file (.map)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.UploadSourcemapResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "422":
          description: only .map files accepted
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload source map
      tags:
      - sourcemaps
  /api/projects/{projectId}/stats:
    get:
      description: 'Returns aggregated error statistics for a project. Used to render
        charts on the dashboard. Three queries (by_date, by_level, totals) run in
        parallel goroutines. Default range: last 30 days.'
      parameters:
      - description: Project UUID
        in: path
        name: projectId
        required: true
        type: string
      - description: 'Start date inclusive (YYYY-MM-DD). Default: 30 days ago'
        in: query
        name: from
        type: string
      - description: 'End date inclusive (YYYY-MM-DD). Default: today'
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.StatsResponse'
        "400":
          description: invalid date format
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Project stats
      tags:
      - stats
swagger: "2.0"